meson: fix pc file generation for uapi
rlib internal deps must not show up in pc file as there are not
c linkable. This is probably a meson nug for rlib handling.
A quick and not so ugly workaround is to tell meson that those rlib
are linked whole (which is actually true, this might be forced for rlib target).