handle: fix type alignment for packed handle
those handle definition need to be packed due to bitfield
handling in C, but, variable of those types must be aligned
on wordsize boundary. This can be done at variable declaration, or,
in type definition.

Be carefull, both __attribute__((packed)) and __attribute__((aligned(x)))
don't apply on the same level. packed is for struct members, and
aligned is for variable of those types.

see: https://gcc.gnu.org/onlinedocs/gcc-3.1.1/gcc/Type-Attributes.html

'''
  This attribute specifies a minimum alignment (in bytes) for variables of the specified type.
  For example, the declarations:
          struct S { short f[3]; } __attribute__ ((aligned (8)));
          typedef int more_aligned_int __attribute__ ((aligned (8)));

  force the compiler to insure (as far as it can) that each variable whose type is struct S
  or more_aligned_int will be allocated and aligned at least on a 8-byte boundary.
'''