task: remove const qualifier
declaring this tab const result in very aggressive optimizations
by compiler that removes all access to this table as it is not
filled at kernel link time but inserted as a blob after task relocation.

In future version, a task may be upgraded, and thus, this table might
be updated. So, this is not really const.

BTW, mpu protection (rx) prevent any writes here in normal operation mode.