workflow: move git safedir workaround
This is not tied to dunamai but a github issue.
user and HOME mismatch while run in container. github action-checkout
can handle safe-directory but this is limited to checkout step and not
the whole job. As we need to use git describe for dynamic versionning,
we have to set the whole workspace as safe.